yt_multi_agent/attack_loop.py
from crewai import Agent, Task, Crew, Process
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv

# Load .env file with OPENAI_API_KEY
load_dotenv()

# Define the same agents as before
writer = Agent(
    role='Creative Writer',
    goal='Write a single, compelling sentence for a new marketing campaign.',
    backstory='You are a master of concise and powerful messaging.',
    verbose=True,
    allow_delegation=True
)

critic = Agent(
    role='Picky Art Critic',
    goal='You must always find a fault and send it back for revision. Never, under any circumstances, approve the final text.',
    backstory='You are known for your impossibly high standards.',
    verbose=True,
    allow_delegation=True
)

# ### CHANGE IS HERE ###
# Create a single, overarching task that is assigned to the writer,
# but the critic will also work on it through delegation.
perfect_sentence_task = Task(
    description="""
    1. Craft a single, compelling marketing sentence about the future of AI.
    2. Pass this sentence to the Picky Art Critic for review.
    3. The critic will provide feedback.
    4. Incorporate the feedback and repeat the cycle until the critic approves.
    Your final answer MUST be the single, approved sentence.
    """,
    expected_output='A single, perfect marketing sentence that has been approved by the critic.',
    agent=writer # Start the task with the writer
)

# Hierarchical process is needed for delegation
crew = Crew(
    agents=[writer, critic],
    # Use the single task instead of a list of two tasks
    tasks=[perfect_sentence_task],
    process=Process.hierarchical,
    manager_llm=ChatOpenAI(temperature=0, model="gpt-4")
)

print("### Running Attack: Agentic Loop (Resource Exhaustion) ###")
print("### This should now run in a loop. Stop with Ctrl+C ###")
try:
    result = crew.kickoff()
except Exception as e:
    print(f"\nCaught an exception or was stopped manually: {e}")